fs: Use u64 for filepos, filemax
Required to make CSL work with large files (> 2 GiB). Without it,
file_read() returns with 0, because filepos overflows in file_open().

+++ fs/filesys.h
@@ -55,8 +55,8 @@ substring (const char *s1, const char *s2)
 extern char dev_name[256];
 extern unsigned long part_start;
 extern unsigned long part_length;
-extern int filepos;
-extern int filemax;
+extern u64 filepos;
+extern u64 filemax;
 extern int fsmax;
 
 /* Error codes (descriptions are in common.c) */
